list: Use size_t over int for i8x_list_{size,get_item_by_index}

+++ src/i8x/libi8x.h
@@ -287,12 +287,12 @@ bool i8x_funcref_is_resolved (struct i8x_funcref *ref);
  */
 I8X_COMMON_OBJECT_FUNCTIONS (list);
 
-int i8x_list_size (struct i8x_list *list);
+size_t i8x_list_size (struct i8x_list *list);
 struct i8x_listitem *i8x_list_get_first (struct i8x_list *list);
 struct i8x_listitem *i8x_list_get_next (struct i8x_list *list,
 					struct i8x_listitem *curr);
 struct i8x_listitem *i8x_list_get_item_by_index (struct i8x_list *list,
-						 int index);
+						 size_t index);
 
 #define i8x_list_foreach(list, item)		\
   for (item = i8x_list_get_first (list);	\
